demo-fastapi-jwt-kafka-prom/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
SQLAlchemy==2.0.36
asyncpg==0.30.0
psycopg2-binary==2.9.9
passlib==1.7.4
PyJWT==2.9.0
pydantic==2.9.2
email-validator==2.2.0        # <-- add this
aiokafka==0.11.0
prometheus-fastapi-instrumentator==6.1.0
python-dotenv==1.0.1
alembic==1.13.3